Zartech2/socialmedia/requirement.txt
Django==2.2.5
django-friendship==1.8.1
django-rest-framework==0.1.0
djangorestframework==3.10.3
pytz==2019.2
sqlparse==0.3.0
